fix issue in crossposting arising from null value not being allowed
